+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## v0.5.2 - 2016-02-21
+- Use webpack watch instead of gulp watching
+- Simplify webpack console logs
+
 ## v0.5.1 - 2016-02-16
 - Improve webpack + gulp build